requirements.txt
dnspython3
SQLAlchemy